implemented keep-option behavior for eraseDir and updated error handling

+++ erase.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"errors"
 	"fmt"
 	"io"
 	"os"
@@ -59,7 +58,10 @@ func eraseDir(path string, op option, errWriter io.Writer) error {
 
 	wg.Wait()
 	if errOccurred.Load() {
-		return errors.New("error occured")
+		return nil
+	}
+	if op.keep {
+		return nil
 	}
 	return os.Remove(path)
 }
